+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   [#22](https://github.com/netsec-ethz/scion-java-client/pull/22)
 - BREAKING CHANGE: `getCurrentPath()` renamed to `getConnectionPath()`
   [#30](https://github.com/netsec-ethz/scion-java-client/pull/30)
+- Cleaned up `MultiMap` utility class 
+  [#34](https://github.com/netsec-ethz/scion-java-client/pull/34)
 
 ### Fixed
 - Fixed: SCMP problem when pinging local AS.